pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "sesyncai"
version = "0.1.0"
description = "Capture, sync, and share your AI project context across tools and machines"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "Pavan Kumar Dhruv", email = "pavankumardhruv@gmail.com" },
]
keywords = ["ai", "context", "sync", "developer-tools", "claude", "cursor", "llm"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
    "typer>=0.9.0",
    "rich>=13.0.0",
    "pyyaml>=6.0",
    "httpx>=0.27.0",
]

[project.scripts]
sesyncai = "sesyncai.cli:app"

[project.urls]
Homepage = "https://github.com/Pavankumardhruv/sesyncai"
Repository = "https://github.com/Pavankumardhruv/sesyncai"
Issues = "https://github.com/Pavankumardhruv/sesyncai/issues"
